Reorganized files
Renamed menu_editor package to user_interface.
Moved the middle end java files to a new source package: middle_end.
These files are not UI related.
The user_interface package is for UI java files.
