+++ Menu_editor/src/user_interface/MenuEditorUI.java
@@ -1,5 +1,5 @@
 
-package menu_editor;
+package user_interface;
 
 import java.sql.*;
 import javax.swing.*;
