+++ Menu_editor/src/user_interface/Menu_UI.java
@@ -1,0 +1,1995 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package user_interface;
+import middle_end.KitchenView;
+import middle_end.Menu;
+import middle_end.MenuItem;
+import middle_end.NewOrder;
+import javax.swing.*;
+/**
+ *
+ * @author Dylan Girard
+ */
+public class Menu_UI extends javax.swing.JFrame {
+
+    /**
+     * Creates new form Menu_UI
+     */
+    public Menu_UI() {
+        initComponents();
+        frame = this;
+        //statspage = new StatsUI();
+        //statspage.setVisible(false);
+    }
+    public MenuEditorUI statspage;
+    public JFrame frame;
+    private KitchenView kitchen;
+    private NewOrder order;
+    private Menu menu;
+    String totalCost = ""+ 0;
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        categoryPanel1 = new javax.swing.JPanel();
+        categoy1Label = new javax.swing.JLabel();
+        category1Button1 = new javax.swing.JButton();
+        category1Button2 = new javax.swing.JButton();
+        category1Button3 = new javax.swing.JButton();
+        category1Button4 = new javax.swing.JButton();
+        category1Button5 = new javax.swing.JButton();
+        category1Button6 = new javax.swing.JButton();
+        category1Button7 = new javax.swing.JButton();
+        category1Button8 = new javax.swing.JButton();
+        placeOrderButton = new javax.swing.JButton();
+        removeItemButton = new javax.swing.JButton();
+        categoryPanel2 = new javax.swing.JPanel();
+        categoyLabel1 = new javax.swing.JLabel();
+        category2Button1 = new javax.swing.JButton();
+        category2Button2 = new javax.swing.JButton();
+        category2Button3 = new javax.swing.JButton();
+        category2Button4 = new javax.swing.JButton();
+        category2Button5 = new javax.swing.JButton();
+        category2Button6 = new javax.swing.JButton();
+        category2Button7 = new javax.swing.JButton();
+        category2Button8 = new javax.swing.JButton();
+        categoryPanel3 = new javax.swing.JPanel();
+        categoyLabel2 = new javax.swing.JLabel();
+        category3Button1 = new javax.swing.JButton();
+        category3Button2 = new javax.swing.JButton();
+        category3Button3 = new javax.swing.JButton();
+        category3Button4 = new javax.swing.JButton();
+        category3Button5 = new javax.swing.JButton();
+        category3Button6 = new javax.swing.JButton();
+        category3Button7 = new javax.swing.JButton();
+        category3Button8 = new javax.swing.JButton();
+        categoryPanel4 = new javax.swing.JPanel();
+        categoyLabel3 = new javax.swing.JLabel();
+        category4Button1 = new javax.swing.JButton();
+        category4Button2 = new javax.swing.JButton();
+        category4Button3 = new javax.swing.JButton();
+        category4Button4 = new javax.swing.JButton();
+        category4Button5 = new javax.swing.JButton();
+        category4Button6 = new javax.swing.JButton();
+        category4Button7 = new javax.swing.JButton();
+        category4Button8 = new javax.swing.JButton();
+        categoryPanel5 = new javax.swing.JPanel();
+        categoyLabel4 = new javax.swing.JLabel();
+        category5Button1 = new javax.swing.JButton();
+        category5Button2 = new javax.swing.JButton();
+        category5Button3 = new javax.swing.JButton();
+        category5Button4 = new javax.swing.JButton();
+        category5Button5 = new javax.swing.JButton();
+        category5Button6 = new javax.swing.JButton();
+        category5Button7 = new javax.swing.JButton();
+        category5Button8 = new javax.swing.JButton();
+        categoryPanel6 = new javax.swing.JPanel();
+        categoyLabel5 = new javax.swing.JLabel();
+        category6Button1 = new javax.swing.JButton();
+        category6Button2 = new javax.swing.JButton();
+        category6Button3 = new javax.swing.JButton();
+        category6Button4 = new javax.swing.JButton();
+        category6Button5 = new javax.swing.JButton();
+        category6Button6 = new javax.swing.JButton();
+        category6Button7 = new javax.swing.JButton();
+        category6Button8 = new javax.swing.JButton();
+        categoryPanel7 = new javax.swing.JPanel();
+        categoyLabel6 = new javax.swing.JLabel();
+        category7Button1 = new javax.swing.JButton();
+        category7Button2 = new javax.swing.JButton();
+        category7Button3 = new javax.swing.JButton();
+        category7Button4 = new javax.swing.JButton();
+        category7Button5 = new javax.swing.JButton();
+        category7Button6 = new javax.swing.JButton();
+        category7Button7 = new javax.swing.JButton();
+        category7Button8 = new javax.swing.JButton();
+        categoryPanel8 = new javax.swing.JPanel();
+        categoyLabel7 = new javax.swing.JLabel();
+        category8Button1 = new javax.swing.JButton();
+        category8Button2 = new javax.swing.JButton();
+        category8Button3 = new javax.swing.JButton();
+        category8Button4 = new javax.swing.JButton();
+        category8Button5 = new javax.swing.JButton();
+        category8Button6 = new javax.swing.JButton();
+        category8Button7 = new javax.swing.JButton();
+        category8Button8 = new javax.swing.JButton();
+        orderTotalLabel = new javax.swing.JLabel();
+        orderTotalLabel1 = new javax.swing.JLabel();
+        jScrollPane2 = new javax.swing.JScrollPane();
+        orderJList = new javax.swing.JList<>();
+
+        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
+        setBackground(new java.awt.Color(0, 0, 0));
+
+        menu = new Menu();
+        menu.initializeCategories(8);
+        kitchen = new KitchenView();
+        order = new NewOrder(kitchen);
+        categoryPanel1.setBackground(new java.awt.Color(0, 0, 0));
+        categoryPanel1.setPreferredSize(new java.awt.Dimension(150, 550));
+
+        String categoryName1 = menu.categoryArray[0].getCategoryName();
+        categoy1Label.setBackground(new java.awt.Color(0, 0, 200));
+        categoy1Label.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        categoy1Label.setForeground(new java.awt.Color(255, 255, 255));
+        categoy1Label.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        categoy1Label.setText(categoryName1);
+
+        //category1Button1.setVisible(true);
+        category1Button1.setBackground(new java.awt.Color(0, 0, 200));
+        menu.categoryArray[0].itemList[0].setItemName("Eggs");
+        String catText11 = menu.categoryArray[0].itemList[0].getItemName();
+        if(catText11.equals(""))
+        {
+            category1Button1.setVisible(false);
+        }
+        else
+        {
+            category1Button1.setVisible(true);
+        }
+        category1Button1.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category1Button1.setForeground(new java.awt.Color(255, 255, 255));
+        category1Button1.setText(catText11);
+        category1Button1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category1Button1ActionPerformed(evt);
+            }
+        });
+
+        //category1Button2.setVisible(true);
+        category1Button2.setBackground(new java.awt.Color(0, 0, 200));
+        category1Button2.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category1Button2.setForeground(new java.awt.Color(255, 255, 255));
+        category1Button2.setText("jButton1");
+        category1Button2.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category1Button2ActionPerformed(evt);
+            }
+        });
+
+        //category1Button3.setVisible(true);
+        category1Button3.setBackground(new java.awt.Color(0, 0, 200));
+        category1Button3.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category1Button3.setForeground(new java.awt.Color(255, 255, 255));
+        category1Button3.setText("jButton1");
+        category1Button3.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category1Button3ActionPerformed(evt);
+            }
+        });
+
+        //category1Button4.setVisible(true);
+        category1Button4.setBackground(new java.awt.Color(0, 0, 200));
+        category1Button4.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category1Button4.setForeground(new java.awt.Color(255, 255, 255));
+        category1Button4.setText("jButton1");
+        category1Button4.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category1Button4ActionPerformed(evt);
+            }
+        });
+
+        //category1Button5.setVisible(true);
+        category1Button5.setBackground(new java.awt.Color(0, 0, 200));
+        category1Button5.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category1Button5.setForeground(new java.awt.Color(255, 255, 255));
+        category1Button5.setText("jButton1");
+        category1Button5.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category1Button5ActionPerformed(evt);
+            }
+        });
+
+        //category1Button6.setVisible(false);
+        category1Button6.setBackground(new java.awt.Color(0, 0, 200));
+        category1Button6.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category1Button6.setForeground(new java.awt.Color(255, 255, 255));
+        category1Button6.setText("jButton1");
+        category1Button6.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category1Button6ActionPerformed(evt);
+            }
+        });
+
+        //category1Button1.setVisible(false);
+        category1Button7.setBackground(new java.awt.Color(0, 0, 200));
+        category1Button7.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category1Button7.setForeground(new java.awt.Color(255, 255, 255));
+        category1Button7.setText("jButton1");
+        category1Button7.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category1Button7ActionPerformed(evt);
+            }
+        });
+
+        //category1Button8.setVisible(false);
+        category1Button8.setBackground(new java.awt.Color(0, 0, 200));
+        category1Button8.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category1Button8.setForeground(new java.awt.Color(255, 255, 255));
+        category1Button8.setText("jButton1");
+        category1Button8.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category1Button8ActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout categoryPanel1Layout = new javax.swing.GroupLayout(categoryPanel1);
+        categoryPanel1.setLayout(categoryPanel1Layout);
+        categoryPanel1Layout.setHorizontalGroup(
+            categoryPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel1Layout.createSequentialGroup()
+                .addGroup(categoryPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(category1Button1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category1Button2, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category1Button3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addGroup(categoryPanel1Layout.createSequentialGroup()
+                        .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(categoy1Label, javax.swing.GroupLayout.PREFERRED_SIZE, 135, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(category1Button4, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category1Button5, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category1Button6, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category1Button7, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category1Button8, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap())
+        );
+        categoryPanel1Layout.setVerticalGroup(
+            categoryPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel1Layout.createSequentialGroup()
+                .addComponent(categoy1Label, javax.swing.GroupLayout.PREFERRED_SIZE, 41, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category1Button1, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category1Button2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category1Button3, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category1Button4, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category1Button5, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category1Button6, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category1Button7, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(category1Button8, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(256, Short.MAX_VALUE))
+        );
+
+        placeOrderButton.setBackground(new java.awt.Color(0, 0, 200));
+        placeOrderButton.setText("Place Order");
+        placeOrderButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                placeOrderButtonActionPerformed(evt);
+            }
+        });
+
+        removeItemButton.setBackground(new java.awt.Color(0, 0, 200));
+        removeItemButton.setText("Remove Item");
+        removeItemButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                removeItemButtonActionPerformed(evt);
+            }
+        });
+
+        categoryPanel2.setBackground(new java.awt.Color(0, 0, 0));
+
+        String categoryName2 = menu.categoryArray[1].getCategoryName();
+        categoyLabel1.setBackground(new java.awt.Color(0, 0, 200));
+        categoyLabel1.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        categoyLabel1.setForeground(new java.awt.Color(255, 255, 255));
+        categoyLabel1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        categoyLabel1.setText(categoryName2);
+
+        //category2Button1.setVisible(true);
+        category2Button1.setBackground(new java.awt.Color(0, 0, 200));
+        category2Button1.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category2Button1.setForeground(new java.awt.Color(255, 255, 255));
+        category2Button1.setText("jButton1");
+        category2Button1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category2Button1ActionPerformed(evt);
+            }
+        });
+
+        //category2Button2.setVisible(false);
+        category2Button2.setBackground(new java.awt.Color(0, 0, 200));
+        category2Button2.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category2Button2.setForeground(new java.awt.Color(255, 255, 255));
+        category2Button2.setText("jButton1");
+        category2Button2.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category2Button2ActionPerformed(evt);
+            }
+        });
+
+        //category2Button3.setVisible(false);
+        category2Button3.setBackground(new java.awt.Color(0, 0, 200));
+        category2Button3.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category2Button3.setForeground(new java.awt.Color(255, 255, 255));
+        category2Button3.setText("jButton1");
+        category2Button3.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category2Button3ActionPerformed(evt);
+            }
+        });
+
+        //category2Button4.setVisible(false);
+        category2Button4.setBackground(new java.awt.Color(0, 0, 200));
+        category2Button4.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category2Button4.setForeground(new java.awt.Color(255, 255, 255));
+        category2Button4.setText("jButton1");
+        category2Button4.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category2Button4ActionPerformed(evt);
+            }
+        });
+
+        //category2Button5.setVisible(false);
+        category2Button5.setBackground(new java.awt.Color(0, 0, 200));
+        category2Button5.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category2Button5.setForeground(new java.awt.Color(255, 255, 255));
+        category2Button5.setText("jButton1");
+        category2Button5.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category2Button5ActionPerformed(evt);
+            }
+        });
+
+        //category2Button6.setVisible(false);
+        category2Button6.setBackground(new java.awt.Color(0, 0, 200));
+        category2Button6.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category2Button6.setForeground(new java.awt.Color(255, 255, 255));
+        category2Button6.setText("jButton1");
+        category2Button6.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category2Button6ActionPerformed(evt);
+            }
+        });
+
+        //category2Button7.setVisible(false);
+        category2Button7.setBackground(new java.awt.Color(0, 0, 200));
+        category2Button7.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category2Button7.setForeground(new java.awt.Color(255, 255, 255));
+        category2Button7.setText("jButton1");
+        category2Button7.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category2Button7ActionPerformed(evt);
+            }
+        });
+
+        //category2Button8.setVisible(false);
+        category2Button8.setBackground(new java.awt.Color(0, 0, 200));
+        category2Button8.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category2Button8.setForeground(new java.awt.Color(255, 255, 255));
+        category2Button8.setText("jButton1");
+        category2Button8.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category2Button8ActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout categoryPanel2Layout = new javax.swing.GroupLayout(categoryPanel2);
+        categoryPanel2.setLayout(categoryPanel2Layout);
+        categoryPanel2Layout.setHorizontalGroup(
+            categoryPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel2Layout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(categoryPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, categoryPanel2Layout.createSequentialGroup()
+                        .addGap(0, 0, Short.MAX_VALUE)
+                        .addComponent(categoyLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 135, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(category2Button1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category2Button2, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category2Button3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category2Button4, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category2Button5, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category2Button6, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category2Button7, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category2Button8, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap())
+        );
+        categoryPanel2Layout.setVerticalGroup(
+            categoryPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel2Layout.createSequentialGroup()
+                .addComponent(categoyLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(category2Button1, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category2Button2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category2Button3, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category2Button4, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category2Button5, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category2Button6, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(8, 8, 8)
+                .addComponent(category2Button7, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category2Button8, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 0, Short.MAX_VALUE))
+        );
+
+        categoryPanel3.setBackground(new java.awt.Color(0, 0, 0));
+        categoryPanel3.setPreferredSize(new java.awt.Dimension(150, 550));
+
+        String categoryName3 = menu.categoryArray[2].getCategoryName();
+        categoyLabel2.setBackground(new java.awt.Color(0, 0, 200));
+        categoyLabel2.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        categoyLabel2.setForeground(new java.awt.Color(255, 255, 255));
+        categoyLabel2.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        categoyLabel2.setText(categoryName3);
+
+        //category3Button1.setVisible(true);
+        category3Button1.setBackground(new java.awt.Color(0, 0, 200));
+        category3Button1.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category3Button1.setForeground(new java.awt.Color(255, 255, 255));
+        category3Button1.setText("jButton1");
+        category3Button1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category3Button1ActionPerformed(evt);
+            }
+        });
+
+        //category3Button2.setVisible(false);
+        category3Button2.setBackground(new java.awt.Color(0, 0, 200));
+        category3Button2.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category3Button2.setForeground(new java.awt.Color(255, 255, 255));
+        category3Button2.setText("jButton1");
+        category3Button2.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category3Button2ActionPerformed(evt);
+            }
+        });
+
+        //category13Button3.setVisible(false);
+        category3Button3.setBackground(new java.awt.Color(0, 0, 200));
+        category3Button3.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category3Button3.setForeground(new java.awt.Color(255, 255, 255));
+        category3Button3.setText("jButton1");
+        category3Button3.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category3Button3ActionPerformed(evt);
+            }
+        });
+
+        //category3Button4.setVisible(false);
+        category3Button4.setBackground(new java.awt.Color(0, 0, 200));
+        category3Button4.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category3Button4.setForeground(new java.awt.Color(255, 255, 255));
+        category3Button4.setText("jButton1");
+        category3Button4.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category3Button4ActionPerformed(evt);
+            }
+        });
+
+        //category3Button5.setVisible(false);
+        category3Button5.setBackground(new java.awt.Color(0, 0, 200));
+        category3Button5.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category3Button5.setForeground(new java.awt.Color(255, 255, 255));
+        category3Button5.setText("jButton1");
+        category3Button5.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category3Button5ActionPerformed(evt);
+            }
+        });
+
+        //category3Button6.setVisible(false);
+        category3Button6.setBackground(new java.awt.Color(0, 0, 200));
+        category3Button6.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category3Button6.setForeground(new java.awt.Color(255, 255, 255));
+        category3Button6.setText("jButton1");
+        category3Button6.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category3Button6ActionPerformed(evt);
+            }
+        });
+
+        //category3Button7.setVisible(false);
+        category3Button7.setBackground(new java.awt.Color(0, 0, 200));
+        category3Button7.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category3Button7.setForeground(new java.awt.Color(255, 255, 255));
+        category3Button7.setText("jButton1");
+        category3Button7.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category3Button7ActionPerformed(evt);
+            }
+        });
+
+        //category3Button8.setVisible(false);
+        category3Button8.setBackground(new java.awt.Color(0, 0, 200));
+        category3Button8.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category3Button8.setForeground(new java.awt.Color(255, 255, 255));
+        category3Button8.setText("jButton1");
+        category3Button8.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category3Button8ActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout categoryPanel3Layout = new javax.swing.GroupLayout(categoryPanel3);
+        categoryPanel3.setLayout(categoryPanel3Layout);
+        categoryPanel3Layout.setHorizontalGroup(
+            categoryPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel3Layout.createSequentialGroup()
+                .addComponent(categoyLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, 135, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 15, Short.MAX_VALUE))
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, categoryPanel3Layout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(categoryPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(category3Button1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category3Button2, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category3Button3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category3Button4, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category3Button5, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category3Button6, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category3Button7, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category3Button8, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap())
+        );
+        categoryPanel3Layout.setVerticalGroup(
+            categoryPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel3Layout.createSequentialGroup()
+                .addComponent(categoyLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(category3Button1, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category3Button2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category3Button3, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category3Button4, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category3Button5, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category3Button6, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(8, 8, 8)
+                .addComponent(category3Button7, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category3Button8, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 0, Short.MAX_VALUE))
+        );
+
+        categoryPanel4.setBackground(new java.awt.Color(0, 0, 0));
+        categoryPanel4.setPreferredSize(new java.awt.Dimension(150, 550));
+
+        String categoryName4 = menu.categoryArray[3].getCategoryName();
+        categoyLabel3.setBackground(new java.awt.Color(0, 0, 200));
+        categoyLabel3.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        categoyLabel3.setForeground(new java.awt.Color(255, 255, 255));
+        categoyLabel3.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        categoyLabel3.setText(categoryName4);
+
+        //category4Button1.setVisible(true);
+        category4Button1.setBackground(new java.awt.Color(0, 0, 200));
+        category4Button1.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category4Button1.setForeground(new java.awt.Color(255, 255, 255));
+        category4Button1.setText("jButton1");
+        category4Button1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category4Button1ActionPerformed(evt);
+            }
+        });
+
+        //category4Button2.setVisible(false);
+        category4Button2.setBackground(new java.awt.Color(0, 0, 200));
+        category4Button2.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category4Button2.setForeground(new java.awt.Color(255, 255, 255));
+        category4Button2.setText("jButton1");
+        category4Button2.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category4Button2ActionPerformed(evt);
+            }
+        });
+
+        //category4Button3.setVisible(false);
+        category4Button3.setBackground(new java.awt.Color(0, 0, 200));
+        category4Button3.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category4Button3.setForeground(new java.awt.Color(255, 255, 255));
+        category4Button3.setText("jButton1");
+        category4Button3.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category4Button3ActionPerformed(evt);
+            }
+        });
+
+        //category4Button4.setVisible(false);
+        category4Button4.setBackground(new java.awt.Color(0, 0, 200));
+        category4Button4.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category4Button4.setForeground(new java.awt.Color(255, 255, 255));
+        category4Button4.setText("jButton1");
+        category4Button4.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category4Button4ActionPerformed(evt);
+            }
+        });
+
+        //category4Button5.setVisible(false);
+        category4Button5.setBackground(new java.awt.Color(0, 0, 200));
+        category4Button5.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category4Button5.setForeground(new java.awt.Color(255, 255, 255));
+        category4Button5.setText("jButton1");
+        category4Button5.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category4Button5ActionPerformed(evt);
+            }
+        });
+
+        //category4Button6.setVisible(false);
+        category4Button6.setBackground(new java.awt.Color(0, 0, 200));
+        category4Button6.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category4Button6.setForeground(new java.awt.Color(255, 255, 255));
+        category4Button6.setText("jButton1");
+        category4Button6.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category4Button6ActionPerformed(evt);
+            }
+        });
+
+        //category4Button7.setVisible(false);
+        category4Button7.setBackground(new java.awt.Color(0, 0, 200));
+        category4Button7.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category4Button7.setForeground(new java.awt.Color(255, 255, 255));
+        category4Button7.setText("jButton1");
+        category4Button7.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category4Button7ActionPerformed(evt);
+            }
+        });
+
+        //category4Button8.setVisible(false);
+        category4Button8.setBackground(new java.awt.Color(0, 0, 200));
+        category4Button8.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category4Button8.setForeground(new java.awt.Color(255, 255, 255));
+        category4Button8.setText("jButton1");
+        category4Button8.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category4Button8ActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout categoryPanel4Layout = new javax.swing.GroupLayout(categoryPanel4);
+        categoryPanel4.setLayout(categoryPanel4Layout);
+        categoryPanel4Layout.setHorizontalGroup(
+            categoryPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel4Layout.createSequentialGroup()
+                .addComponent(categoyLabel3, javax.swing.GroupLayout.PREFERRED_SIZE, 135, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 15, Short.MAX_VALUE))
+            .addGroup(categoryPanel4Layout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(categoryPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(category4Button1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category4Button2, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category4Button3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category4Button4, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category4Button5, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category4Button6, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category4Button7, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category4Button8, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap())
+        );
+        categoryPanel4Layout.setVerticalGroup(
+            categoryPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel4Layout.createSequentialGroup()
+                .addComponent(categoyLabel3, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(category4Button1, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category4Button2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category4Button3, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category4Button4, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category4Button5, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category4Button6, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(8, 8, 8)
+                .addComponent(category4Button7, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category4Button8, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 0, Short.MAX_VALUE))
+        );
+
+        categoryPanel5.setBackground(new java.awt.Color(0, 0, 0));
+        categoryPanel5.setPreferredSize(new java.awt.Dimension(150, 550));
+
+        String categoryName5 = menu.categoryArray[4].getCategoryName();
+        categoyLabel4.setBackground(new java.awt.Color(0, 0, 200));
+        categoyLabel4.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        categoyLabel4.setForeground(new java.awt.Color(255, 255, 255));
+        categoyLabel4.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        categoyLabel4.setText(categoryName5);
+
+        //category5Button1.setVisible(true);
+        category5Button1.setBackground(new java.awt.Color(0, 0, 200));
+        category5Button1.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category5Button1.setForeground(new java.awt.Color(255, 255, 255));
+        category5Button1.setText("jButton1");
+        category5Button1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category5Button1ActionPerformed(evt);
+            }
+        });
+
+        //category5Button2.setVisible(false);
+        category5Button2.setBackground(new java.awt.Color(0, 0, 200));
+        category5Button2.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category5Button2.setForeground(new java.awt.Color(255, 255, 255));
+        category5Button2.setText("jButton1");
+        category5Button2.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category5Button2ActionPerformed(evt);
+            }
+        });
+
+        //category5Button3.setVisible(false);
+        category5Button3.setBackground(new java.awt.Color(0, 0, 200));
+        category5Button3.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category5Button3.setForeground(new java.awt.Color(255, 255, 255));
+        category5Button3.setText("jButton1");
+        category5Button3.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category5Button3ActionPerformed(evt);
+            }
+        });
+
+        //category5Button4.setVisible(false);
+        category5Button4.setBackground(new java.awt.Color(0, 0, 200));
+        category5Button4.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category5Button4.setForeground(new java.awt.Color(255, 255, 255));
+        category5Button4.setText("jButton1");
+        category5Button4.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category5Button4ActionPerformed(evt);
+            }
+        });
+
+        //category5Button5.setVisible(false);
+        category5Button5.setBackground(new java.awt.Color(0, 0, 200));
+        category5Button5.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category5Button5.setForeground(new java.awt.Color(255, 255, 255));
+        category5Button5.setText("jButton1");
+        category5Button5.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category5Button5ActionPerformed(evt);
+            }
+        });
+
+        //category5Button6.setVisible(false);
+        category5Button6.setBackground(new java.awt.Color(0, 0, 200));
+        category5Button6.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category5Button6.setForeground(new java.awt.Color(255, 255, 255));
+        category5Button6.setText("jButton1");
+        category5Button6.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category5Button6ActionPerformed(evt);
+            }
+        });
+
+        //category5Button7.setVisible(false);
+        category5Button7.setBackground(new java.awt.Color(0, 0, 200));
+        category5Button7.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category5Button7.setForeground(new java.awt.Color(255, 255, 255));
+        category5Button7.setText("jButton1");
+        category5Button7.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category5Button7ActionPerformed(evt);
+            }
+        });
+
+        //category5Button8.setVisible(false);
+        category5Button8.setBackground(new java.awt.Color(0, 0, 200));
+        category5Button8.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category5Button8.setForeground(new java.awt.Color(255, 255, 255));
+        category5Button8.setText("jButton1");
+        category5Button8.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category5Button8ActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout categoryPanel5Layout = new javax.swing.GroupLayout(categoryPanel5);
+        categoryPanel5.setLayout(categoryPanel5Layout);
+        categoryPanel5Layout.setHorizontalGroup(
+            categoryPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel5Layout.createSequentialGroup()
+                .addGroup(categoryPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addGroup(categoryPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(category5Button1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(category5Button2, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(category5Button3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(category5Button4, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(category5Button5, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(category5Button6, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(category5Button7, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(category5Button8, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                    .addComponent(categoyLabel4, javax.swing.GroupLayout.PREFERRED_SIZE, 135, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addGap(0, 15, Short.MAX_VALUE))
+        );
+        categoryPanel5Layout.setVerticalGroup(
+            categoryPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel5Layout.createSequentialGroup()
+                .addComponent(categoyLabel4, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(category5Button1, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category5Button2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category5Button3, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category5Button4, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category5Button5, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category5Button6, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(8, 8, 8)
+                .addComponent(category5Button7, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category5Button8, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 247, Short.MAX_VALUE))
+        );
+
+        categoryPanel6.setBackground(new java.awt.Color(0, 0, 0));
+        categoryPanel6.setPreferredSize(new java.awt.Dimension(150, 550));
+
+        String categoryName6 = menu.categoryArray[5].getCategoryName();
+        categoyLabel5.setBackground(new java.awt.Color(0, 0, 200));
+        categoyLabel5.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        categoyLabel5.setForeground(new java.awt.Color(255, 255, 255));
+        categoyLabel5.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        categoyLabel5.setText(categoryName6);
+
+        //category6Button1.setVisible(true);
+        category6Button1.setBackground(new java.awt.Color(0, 0, 200));
+        category6Button1.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category6Button1.setForeground(new java.awt.Color(255, 255, 255));
+        category6Button1.setText("jButton1");
+        category6Button1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category6Button1ActionPerformed(evt);
+            }
+        });
+
+        //category6Button2.setVisible(false);
+        category6Button2.setBackground(new java.awt.Color(0, 0, 200));
+        category6Button2.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category6Button2.setForeground(new java.awt.Color(255, 255, 255));
+        category6Button2.setText("jButton1");
+        category6Button2.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category6Button2ActionPerformed(evt);
+            }
+        });
+
+        //category6Button3.setVisible(false);
+        category6Button3.setBackground(new java.awt.Color(0, 0, 200));
+        category6Button3.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category6Button3.setForeground(new java.awt.Color(255, 255, 255));
+        category6Button3.setText("jButton1");
+        category6Button3.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category6Button3ActionPerformed(evt);
+            }
+        });
+
+        //category6Button4.setVisible(false);
+        category6Button4.setBackground(new java.awt.Color(0, 0, 200));
+        category6Button4.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category6Button4.setForeground(new java.awt.Color(255, 255, 255));
+        category6Button4.setText("jButton1");
+        category6Button4.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category6Button4ActionPerformed(evt);
+            }
+        });
+
+        //category6Button5.setVisible(false);
+        category6Button5.setBackground(new java.awt.Color(0, 0, 200));
+        category6Button5.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category6Button5.setForeground(new java.awt.Color(255, 255, 255));
+        category6Button5.setText("jButton1");
+        category6Button5.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category6Button5ActionPerformed(evt);
+            }
+        });
+
+        //category6Button6.setVisible(false);
+        category6Button6.setBackground(new java.awt.Color(0, 0, 200));
+        category6Button6.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category6Button6.setForeground(new java.awt.Color(255, 255, 255));
+        category6Button6.setText("jButton1");
+        category6Button6.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category6Button6ActionPerformed(evt);
+            }
+        });
+
+        //category6Button7.setVisible(false);
+        category6Button7.setBackground(new java.awt.Color(0, 0, 200));
+        category6Button7.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category6Button7.setForeground(new java.awt.Color(255, 255, 255));
+        category6Button7.setText("jButton1");
+        category6Button7.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category6Button7ActionPerformed(evt);
+            }
+        });
+
+        //category6Button8.setVisible(false);
+        category6Button8.setBackground(new java.awt.Color(0, 0, 200));
+        category6Button8.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category6Button8.setForeground(new java.awt.Color(255, 255, 255));
+        category6Button8.setText("jButton1");
+        category6Button8.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category6Button8ActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout categoryPanel6Layout = new javax.swing.GroupLayout(categoryPanel6);
+        categoryPanel6.setLayout(categoryPanel6Layout);
+        categoryPanel6Layout.setHorizontalGroup(
+            categoryPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel6Layout.createSequentialGroup()
+                .addComponent(categoyLabel5, javax.swing.GroupLayout.PREFERRED_SIZE, 135, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 15, Short.MAX_VALUE))
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, categoryPanel6Layout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(categoryPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(category6Button1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category6Button2, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category6Button3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category6Button4, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category6Button5, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category6Button6, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category6Button7, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category6Button8, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap())
+        );
+        categoryPanel6Layout.setVerticalGroup(
+            categoryPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel6Layout.createSequentialGroup()
+                .addComponent(categoyLabel5, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(category6Button1, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category6Button2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category6Button3, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category6Button4, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category6Button5, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category6Button6, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(8, 8, 8)
+                .addComponent(category6Button7, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category6Button8, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 0, Short.MAX_VALUE))
+        );
+
+        categoryPanel7.setBackground(new java.awt.Color(0, 0, 0));
+        categoryPanel7.setPreferredSize(new java.awt.Dimension(150, 550));
+
+        String categoryName7 = menu.categoryArray[6].getCategoryName();
+        categoyLabel6.setBackground(new java.awt.Color(0, 0, 200));
+        categoyLabel6.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        categoyLabel6.setForeground(new java.awt.Color(255, 255, 255));
+        categoyLabel6.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        categoyLabel6.setText(categoryName7);
+
+        //category7Button1.setVisible(true);
+        category7Button1.setBackground(new java.awt.Color(0, 0, 200));
+        category7Button1.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category7Button1.setForeground(new java.awt.Color(255, 255, 255));
+        category7Button1.setText("jButton1");
+        category7Button1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category7Button1ActionPerformed(evt);
+            }
+        });
+
+        //category7Button2.setVisible(false);
+        category7Button2.setBackground(new java.awt.Color(0, 0, 200));
+        category7Button2.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category7Button2.setForeground(new java.awt.Color(255, 255, 255));
+        category7Button2.setText("jButton1");
+        category7Button2.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category7Button2ActionPerformed(evt);
+            }
+        });
+
+        //category7Button3.setVisible(false);
+        category7Button3.setBackground(new java.awt.Color(0, 0, 200));
+        category7Button3.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category7Button3.setForeground(new java.awt.Color(255, 255, 255));
+        category7Button3.setText("jButton1");
+        category7Button3.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category7Button3ActionPerformed(evt);
+            }
+        });
+
+        //category7Button4.setVisible(false);
+        category7Button4.setBackground(new java.awt.Color(0, 0, 200));
+        category7Button4.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category7Button4.setForeground(new java.awt.Color(255, 255, 255));
+        category7Button4.setText("jButton1");
+        category7Button4.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category7Button4ActionPerformed(evt);
+            }
+        });
+
+        //category7Button5.setVisible(false);
+        category7Button5.setBackground(new java.awt.Color(0, 0, 200));
+        category7Button5.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category7Button5.setForeground(new java.awt.Color(255, 255, 255));
+        category7Button5.setText("jButton1");
+        category7Button5.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category7Button5ActionPerformed(evt);
+            }
+        });
+
+        //ategory7Button6.setVisible(false);
+        category7Button6.setBackground(new java.awt.Color(0, 0, 200));
+        category7Button6.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category7Button6.setForeground(new java.awt.Color(255, 255, 255));
+        category7Button6.setText("jButton1");
+        category7Button6.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category7Button6ActionPerformed(evt);
+            }
+        });
+
+        //category7Button7.setVisible(false);
+        category7Button7.setBackground(new java.awt.Color(0, 0, 200));
+        category7Button7.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category7Button7.setForeground(new java.awt.Color(255, 255, 255));
+        category7Button7.setText("jButton1");
+        category7Button7.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category7Button7ActionPerformed(evt);
+            }
+        });
+
+        //category7Button8.setVisible(false);
+        category7Button8.setBackground(new java.awt.Color(0, 0, 200));
+        category7Button8.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category7Button8.setForeground(new java.awt.Color(255, 255, 255));
+        category7Button8.setText("jButton1");
+        category7Button8.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category7Button8ActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout categoryPanel7Layout = new javax.swing.GroupLayout(categoryPanel7);
+        categoryPanel7.setLayout(categoryPanel7Layout);
+        categoryPanel7Layout.setHorizontalGroup(
+            categoryPanel7Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel7Layout.createSequentialGroup()
+                .addComponent(categoyLabel6, javax.swing.GroupLayout.PREFERRED_SIZE, 135, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 15, Short.MAX_VALUE))
+            .addGroup(categoryPanel7Layout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(categoryPanel7Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(category7Button1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category7Button2, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category7Button3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category7Button4, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category7Button5, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category7Button6, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category7Button7, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category7Button8, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap())
+        );
+        categoryPanel7Layout.setVerticalGroup(
+            categoryPanel7Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel7Layout.createSequentialGroup()
+                .addComponent(categoyLabel6, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(category7Button1, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category7Button2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category7Button3, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category7Button4, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category7Button5, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category7Button6, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(8, 8, 8)
+                .addComponent(category7Button7, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category7Button8, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 0, Short.MAX_VALUE))
+        );
+
+        categoryPanel8.setBackground(new java.awt.Color(0, 0, 0));
+        categoryPanel8.setPreferredSize(new java.awt.Dimension(150, 550));
+
+        String categoryName8 = menu.categoryArray[7].getCategoryName();
+        categoyLabel7.setBackground(new java.awt.Color(0, 0, 200));
+        categoyLabel7.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        categoyLabel7.setForeground(new java.awt.Color(255, 255, 255));
+        categoyLabel7.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        categoyLabel7.setText(categoryName8);
+
+        //category8Button1.setVisible(true);
+        category8Button1.setBackground(new java.awt.Color(0, 0, 200));
+        category8Button1.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category8Button1.setForeground(new java.awt.Color(255, 255, 255));
+        category8Button1.setText("jButton1");
+        category8Button1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category8Button1ActionPerformed(evt);
+            }
+        });
+
+        //category8Button2.setVisible(false);
+        category8Button2.setBackground(new java.awt.Color(0, 0, 200));
+        category8Button2.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category8Button2.setForeground(new java.awt.Color(255, 255, 255));
+        category8Button2.setText("jButton1");
+        category8Button2.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category8Button2ActionPerformed(evt);
+            }
+        });
+
+        //category8Button3.setVisible(false);
+        category8Button3.setBackground(new java.awt.Color(0, 0, 200));
+        category8Button3.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category8Button3.setForeground(new java.awt.Color(255, 255, 255));
+        category8Button3.setText("jButton1");
+        category8Button3.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category8Button3ActionPerformed(evt);
+            }
+        });
+
+        //category8Button4.setVisible(false);
+        category8Button4.setBackground(new java.awt.Color(0, 0, 200));
+        category8Button4.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category8Button4.setForeground(new java.awt.Color(255, 255, 255));
+        category8Button4.setText("jButton1");
+        category8Button4.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category8Button4ActionPerformed(evt);
+            }
+        });
+
+        //category8Button5.setVisible(false);
+        category8Button5.setBackground(new java.awt.Color(0, 0, 200));
+        category8Button5.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category8Button5.setForeground(new java.awt.Color(255, 255, 255));
+        category8Button5.setText("jButton1");
+        category8Button5.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category8Button5ActionPerformed(evt);
+            }
+        });
+
+        //category8Button6.setVisible(false);
+        category8Button6.setBackground(new java.awt.Color(0, 0, 200));
+        category8Button6.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category8Button6.setForeground(new java.awt.Color(255, 255, 255));
+        category8Button6.setText("jButton1");
+        category8Button6.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category8Button6ActionPerformed(evt);
+            }
+        });
+
+        //category8Button7.setVisible(false);
+        category8Button7.setBackground(new java.awt.Color(0, 0, 200));
+        category8Button7.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category8Button7.setForeground(new java.awt.Color(255, 255, 255));
+        category8Button7.setText("jButton1");
+        category8Button7.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category8Button7ActionPerformed(evt);
+            }
+        });
+
+        //category8Button8.setVisible(false);
+        category8Button8.setBackground(new java.awt.Color(0, 0, 200));
+        category8Button8.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        category8Button8.setForeground(new java.awt.Color(255, 255, 255));
+        category8Button8.setText("jButton1");
+        category8Button8.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                category8Button8ActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout categoryPanel8Layout = new javax.swing.GroupLayout(categoryPanel8);
+        categoryPanel8.setLayout(categoryPanel8Layout);
+        categoryPanel8Layout.setHorizontalGroup(
+            categoryPanel8Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel8Layout.createSequentialGroup()
+                .addGroup(categoryPanel8Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addComponent(category8Button1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(categoyLabel7, javax.swing.GroupLayout.PREFERRED_SIZE, 135, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(category8Button2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category8Button3, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category8Button4, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category8Button5, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category8Button6, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category8Button7, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(category8Button8, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addGap(0, 15, Short.MAX_VALUE))
+        );
+        categoryPanel8Layout.setVerticalGroup(
+            categoryPanel8Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(categoryPanel8Layout.createSequentialGroup()
+                .addComponent(categoyLabel7, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(category8Button1, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category8Button2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category8Button3, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category8Button4, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category8Button5, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category8Button6, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category8Button7, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(category8Button8, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 249, Short.MAX_VALUE))
+        );
+
+        orderTotalLabel.setBackground(new java.awt.Color(0, 0, 0));
+        orderTotalLabel.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        orderTotalLabel.setText("Total Cost:       $");
+
+        orderTotalLabel1.setBackground(new java.awt.Color(0, 0, 0));
+        orderTotalLabel1.setFont(new java.awt.Font("Vani", 3, 14)); // NOI18N
+        orderTotalLabel1.setText(totalCost);
+
+        DefaultListModel defaultModel = new DefaultListModel();
+        orderJList.setModel(defaultModel);
+        jScrollPane2.setViewportView(orderJList);
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
+        getContentPane().setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(categoryPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, 155, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(categoryPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(categoryPanel3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(categoryPanel4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(categoryPanel5, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(categoryPanel6, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(categoryPanel7, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(categoryPanel8, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(layout.createSequentialGroup()
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(removeItemButton, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                            .addComponent(placeOrderButton, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                        .addContainerGap())
+                    .addGroup(layout.createSequentialGroup()
+                        .addComponent(orderTotalLabel, javax.swing.GroupLayout.PREFERRED_SIZE, 117, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(orderTotalLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 117, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(0, 10, Short.MAX_VALUE))
+                    .addComponent(jScrollPane2, javax.swing.GroupLayout.PREFERRED_SIZE, 0, Short.MAX_VALUE)))
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(categoryPanel5, javax.swing.GroupLayout.DEFAULT_SIZE, 739, Short.MAX_VALUE)
+                    .addGroup(layout.createSequentialGroup()
+                        .addComponent(jScrollPane2, javax.swing.GroupLayout.PREFERRED_SIZE, 573, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(orderTotalLabel, javax.swing.GroupLayout.PREFERRED_SIZE, 52, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(orderTotalLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 52, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(placeOrderButton, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                        .addComponent(removeItemButton, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(0, 22, Short.MAX_VALUE))
+                    .addComponent(categoryPanel8, javax.swing.GroupLayout.DEFAULT_SIZE, 739, Short.MAX_VALUE))
+                .addContainerGap())
+            .addComponent(categoryPanel4, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, 750, Short.MAX_VALUE)
+            .addComponent(categoryPanel3, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, 750, Short.MAX_VALUE)
+            .addComponent(categoryPanel2, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(categoryPanel1, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, 750, Short.MAX_VALUE)
+            .addComponent(categoryPanel6, javax.swing.GroupLayout.DEFAULT_SIZE, 750, Short.MAX_VALUE)
+            .addComponent(categoryPanel7, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, 750, Short.MAX_VALUE)
+        );
+
+        pack();
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void category7Button6ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category7Button6ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[6].itemList[5];
+        JButton button = category7Button6;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category7Button6ActionPerformed
+
+    private void removeItemButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_removeItemButtonActionPerformed
+    //Prevents items from being removed when no list items are being selected.
+     if(!orderJList.isSelectionEmpty())
+     {
+         DefaultListModel model = getListModel();
+
+         //Remove the item from the list.
+         MenuItem item = order.removeFromOrder(orderJList.getSelectedIndex());
+         model.removeElement(orderJList.getSelectedValue()); 
+         // item.setItemPrice(75);
+         order.removeFromTotal(item.getItemPrice());
+         //Safeguard against having a negative totalCost
+         if(order.getTotalCost() < 0)
+         {
+             order.addToTotal(-order.getTotalCost());
+             totalCost = "" + 0;
+         }
+         else
+         {
+            totalCost = "" + order.getTotalCost();
+         }
+
+         orderTotalLabel1.setText(totalCost);
+
+     }
+    }//GEN-LAST:event_removeItemButtonActionPerformed
+
+    private void category1Button1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category1Button1ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[0].itemList[0];
+        JButton button = category1Button1;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category1Button1ActionPerformed
+
+    private void category1Button2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category1Button2ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[0].itemList[1];
+        JButton button = category1Button2;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category1Button2ActionPerformed
+
+    private void placeOrderButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_placeOrderButtonActionPerformed
+        //Sends the order to the kitchen by adding the order to the list of orders for the kitchen to see.
+        order.sendToKitchen();
+        JOptionPane.showMessageDialog(null, "Order was sent to the kitchen.");
+        //frame.setVisible(false);
+        //statspage.setVisible(true);
+        //System.exit(0);
+        
+        
+    }//GEN-LAST:event_placeOrderButtonActionPerformed
+
+    private void category1Button3ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category1Button3ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[0].itemList[2];
+        JButton button = category1Button3;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category1Button3ActionPerformed
+
+    private void category1Button4ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category1Button4ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[0].itemList[3];
+        JButton button = category1Button4;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category1Button4ActionPerformed
+
+    private void category1Button5ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category1Button5ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[0].itemList[4];
+        JButton button = category1Button5;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category1Button5ActionPerformed
+
+    private void category1Button6ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category1Button6ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[0].itemList[5];
+        JButton button = category1Button6;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category1Button6ActionPerformed
+
+    private void category1Button7ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category1Button7ActionPerformed
+    DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[0].itemList[6];
+        JButton button = category1Button7;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category1Button7ActionPerformed
+
+    private void category1Button8ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category1Button8ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[0].itemList[7];
+        JButton button = category1Button8;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category1Button8ActionPerformed
+
+    private void category2Button1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category2Button1ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[1].itemList[0];
+        JButton button = category2Button1;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category2Button1ActionPerformed
+
+    private void category2Button2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category2Button2ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[1].itemList[1];
+        JButton button = category2Button2;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category2Button2ActionPerformed
+
+    private void category2Button3ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category2Button3ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[1].itemList[2];
+        JButton button = category2Button3;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category2Button3ActionPerformed
+
+    private void category2Button4ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category2Button4ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[1].itemList[3];
+        JButton button = category2Button4;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category2Button4ActionPerformed
+
+    private void category2Button5ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category2Button5ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[1].itemList[4];
+        JButton button = category2Button5;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category2Button5ActionPerformed
+
+    private void category2Button6ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category2Button6ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[1].itemList[5];
+        JButton button = category2Button6;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category2Button6ActionPerformed
+
+    private void category2Button7ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category2Button7ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[1].itemList[6];
+        JButton button = category2Button7;
+        addItemToOrder(item, model, button);    }//GEN-LAST:event_category2Button7ActionPerformed
+
+    private void category2Button8ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category2Button8ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[1].itemList[7];
+        JButton button = category2Button8;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category2Button8ActionPerformed
+
+    private void category3Button1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category3Button1ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[2].itemList[0];
+        JButton button = category3Button1;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category3Button1ActionPerformed
+
+    private void category3Button2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category3Button2ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[2].itemList[1];
+        JButton button = category3Button2;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category3Button2ActionPerformed
+
+    private void category3Button3ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category3Button3ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[2].itemList[2];
+        JButton button = category3Button3;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category3Button3ActionPerformed
+
+    private void category3Button4ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category3Button4ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[2].itemList[3];
+        JButton button = category3Button4;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category3Button4ActionPerformed
+
+    private void category3Button5ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category3Button5ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[2].itemList[4];
+        JButton button = category3Button5;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category3Button5ActionPerformed
+
+    private void category3Button6ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category3Button6ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[2].itemList[5];
+        JButton button = category3Button6;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category3Button6ActionPerformed
+
+    private void category3Button7ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category3Button7ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[2].itemList[6];
+        JButton button = category3Button7;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category3Button7ActionPerformed
+
+    private void category3Button8ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category3Button8ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[2].itemList[7];
+        JButton button = category3Button8;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category3Button8ActionPerformed
+
+    private void category4Button1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category4Button1ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[3].itemList[0];
+        JButton button = category4Button1;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category4Button1ActionPerformed
+
+    private void category4Button2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category4Button2ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[3].itemList[1];
+        JButton button = category4Button2;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category4Button2ActionPerformed
+
+    private void category4Button3ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category4Button3ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[3].itemList[2];
+        JButton button = category4Button3;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category4Button3ActionPerformed
+
+    private void category4Button4ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category4Button4ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[3].itemList[3];
+        JButton button = category4Button4;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category4Button4ActionPerformed
+
+    private void category4Button5ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category4Button5ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[3].itemList[4];
+        JButton button = category4Button5;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category4Button5ActionPerformed
+
+    private void category4Button6ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category4Button6ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[3].itemList[5];
+        JButton button = category4Button6;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category4Button6ActionPerformed
+
+    private void category4Button7ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category4Button7ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[3].itemList[6];
+        JButton button = category4Button7;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category4Button7ActionPerformed
+
+    private void category4Button8ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category4Button8ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[3].itemList[7];
+        JButton button = category4Button8;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category4Button8ActionPerformed
+
+    private void category5Button1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category5Button1ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[4].itemList[0];
+        JButton button = category5Button1;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category5Button1ActionPerformed
+
+    private void category5Button2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category5Button2ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[4].itemList[1];
+        JButton button = category5Button2;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category5Button2ActionPerformed
+
+    private void category5Button3ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category5Button3ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[4].itemList[2];
+        JButton button = category5Button3;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category5Button3ActionPerformed
+
+    private void category5Button4ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category5Button4ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[4].itemList[3];
+        JButton button = category5Button4;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category5Button4ActionPerformed
+
+    private void category5Button5ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category5Button5ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[4].itemList[4];
+        JButton button = category5Button5;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category5Button5ActionPerformed
+
+    private void category5Button6ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category5Button6ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[4].itemList[5];
+        JButton button = category5Button6;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category5Button6ActionPerformed
+
+    private void category5Button7ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category5Button7ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[4].itemList[6];
+        JButton button = category5Button7;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category5Button7ActionPerformed
+
+    private void category5Button8ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category5Button8ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[4].itemList[7];
+        JButton button = category5Button8;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category5Button8ActionPerformed
+
+    private void category6Button1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category6Button1ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[5].itemList[0];
+        JButton button = category6Button1;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category6Button1ActionPerformed
+
+    private void category6Button2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category6Button2ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[5].itemList[1];
+        JButton button = category6Button2;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category6Button2ActionPerformed
+
+    private void category6Button3ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category6Button3ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[5].itemList[2];
+        JButton button = category6Button3;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category6Button3ActionPerformed
+
+    private void category6Button4ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category6Button4ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[5].itemList[3];
+        JButton button = category6Button4;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category6Button4ActionPerformed
+
+    private void category6Button5ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category6Button5ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[5].itemList[4];
+        JButton button = category6Button5;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category6Button5ActionPerformed
+
+    private void category6Button6ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category6Button6ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[5].itemList[5];
+        JButton button = category6Button6;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category6Button6ActionPerformed
+
+    private void category6Button7ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category6Button7ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[5].itemList[6];
+        JButton button = category6Button7;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category6Button7ActionPerformed
+
+    private void category6Button8ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category6Button8ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[5].itemList[7];
+        JButton button = category6Button8;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category6Button8ActionPerformed
+
+    private void category7Button1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category7Button1ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[6].itemList[0];
+        JButton button = category7Button1;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category7Button1ActionPerformed
+
+    private void category7Button2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category7Button2ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[6].itemList[1];
+        JButton button = category7Button2;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category7Button2ActionPerformed
+
+    private void category7Button3ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category7Button3ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[6].itemList[2];
+        JButton button = category7Button3;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category7Button3ActionPerformed
+
+    private void category7Button4ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category7Button4ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[6].itemList[3];
+        JButton button = category7Button4;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category7Button4ActionPerformed
+
+    private void category7Button5ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category7Button5ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[6].itemList[4];
+        JButton button = category7Button5;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category7Button5ActionPerformed
+
+    private void category7Button7ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category7Button7ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[6].itemList[6];
+        JButton button = category7Button7;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category7Button7ActionPerformed
+
+    private void category7Button8ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category7Button8ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[6].itemList[7];
+        JButton button = category7Button8;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category7Button8ActionPerformed
+
+    private void category8Button1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category8Button1ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[7].itemList[0];
+        JButton button = category8Button1;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category8Button1ActionPerformed
+
+    private void category8Button2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category8Button2ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[7].itemList[1];
+        JButton button = category8Button2;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category8Button2ActionPerformed
+
+    private void category8Button3ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category8Button3ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[7].itemList[2];
+        JButton button = category8Button3;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category8Button3ActionPerformed
+
+    private void category8Button4ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category8Button4ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[7].itemList[3];
+        JButton button = category8Button4;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category8Button4ActionPerformed
+
+    private void category8Button5ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category8Button5ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[7].itemList[4];
+        JButton button = category8Button5;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category8Button5ActionPerformed
+
+    private void category8Button6ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category8Button6ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[7].itemList[5];
+        JButton button = category8Button6;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category8Button6ActionPerformed
+
+    private void category8Button7ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category8Button7ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[7].itemList[6];
+        JButton button = category8Button7;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category8Button7ActionPerformed
+
+    private void category8Button8ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_category8Button8ActionPerformed
+        DefaultListModel model = getListModel();
+        MenuItem item = menu.categoryArray[7].itemList[7];
+        JButton button = category8Button8;
+        addItemToOrder(item, model, button);
+    }//GEN-LAST:event_category8Button8ActionPerformed
+
+    //Function to get the list model of the orderJList
+    private DefaultListModel getListModel()
+    {
+         DefaultListModel model = (DefaultListModel) orderJList.getModel();
+       if(model == null)
+       {
+           model = new DefaultListModel();
+           orderJList.setModel(model);
+       }
+        return model;
+    }
+    //Function to add items to the orderJList and to the order itself when the user 
+    //clicks on a button.
+    private void addItemToOrder(MenuItem item, DefaultListModel model, JButton button)
+    {
+        //Add a comment to the item in the order. For example, if a customer orders a 
+       //cheeseburger, maybe they want no lettuce or extra ketchup.
+       String comment = JOptionPane.showInputDialog(null, "Any special instructions regarding your order?\n"
+               + "Leave the text field blank if there are no special instructions.");
+
+       //Add the item along with any special instructions to the list.
+       model.addElement(button.getText() + " " + comment);
+       item.setItemPrice(100);
+       order.addToTotal(item.getItemPrice());
+       totalCost = "" + order.getTotalCost();
+       orderTotalLabel1.setText(totalCost);
+       order.addToOrder(item);
+    }
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String args[]) {
+        /* Set the Nimbus look and feel */
+        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
+        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
+         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
+         */
+        try {
+            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
+                if ("Nimbus".equals(info.getName())) {
+                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
+                    break;
+                }
+            }
+        } catch (ClassNotFoundException ex) {
+            java.util.logging.Logger.getLogger(Menu_UI.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (InstantiationException ex) {
+            java.util.logging.Logger.getLogger(Menu_UI.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (IllegalAccessException ex) {
+            java.util.logging.Logger.getLogger(Menu_UI.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
+            java.util.logging.Logger.getLogger(Menu_UI.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        }
+        //</editor-fold>
+
+        /* Create and display the form */
+        java.awt.EventQueue.invokeLater(new Runnable() {
+            public void run() {
+                new Menu_UI().setVisible(true);
+            }
+        });
+    }
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JButton category1Button1;
+    private javax.swing.JButton category1Button2;
+    private javax.swing.JButton category1Button3;
+    private javax.swing.JButton category1Button4;
+    private javax.swing.JButton category1Button5;
+    private javax.swing.JButton category1Button6;
+    private javax.swing.JButton category1Button7;
+    private javax.swing.JButton category1Button8;
+    private javax.swing.JButton category2Button1;
+    private javax.swing.JButton category2Button2;
+    private javax.swing.JButton category2Button3;
+    private javax.swing.JButton category2Button4;
+    private javax.swing.JButton category2Button5;
+    private javax.swing.JButton category2Button6;
+    private javax.swing.JButton category2Button7;
+    private javax.swing.JButton category2Button8;
+    private javax.swing.JButton category3Button1;
+    private javax.swing.JButton category3Button2;
+    private javax.swing.JButton category3Button3;
+    private javax.swing.JButton category3Button4;
+    private javax.swing.JButton category3Button5;
+    private javax.swing.JButton category3Button6;
+    private javax.swing.JButton category3Button7;
+    private javax.swing.JButton category3Button8;
+    private javax.swing.JButton category4Button1;
+    private javax.swing.JButton category4Button2;
+    private javax.swing.JButton category4Button3;
+    private javax.swing.JButton category4Button4;
+    private javax.swing.JButton category4Button5;
+    private javax.swing.JButton category4Button6;
+    private javax.swing.JButton category4Button7;
+    private javax.swing.JButton category4Button8;
+    private javax.swing.JButton category5Button1;
+    private javax.swing.JButton category5Button2;
+    private javax.swing.JButton category5Button3;
+    private javax.swing.JButton category5Button4;
+    private javax.swing.JButton category5Button5;
+    private javax.swing.JButton category5Button6;
+    private javax.swing.JButton category5Button7;
+    private javax.swing.JButton category5Button8;
+    private javax.swing.JButton category6Button1;
+    private javax.swing.JButton category6Button2;
+    private javax.swing.JButton category6Button3;
+    private javax.swing.JButton category6Button4;
+    private javax.swing.JButton category6Button5;
+    private javax.swing.JButton category6Button6;
+    private javax.swing.JButton category6Button7;
+    private javax.swing.JButton category6Button8;
+    private javax.swing.JButton category7Button1;
+    private javax.swing.JButton category7Button2;
+    private javax.swing.JButton category7Button3;
+    private javax.swing.JButton category7Button4;
+    private javax.swing.JButton category7Button5;
+    private javax.swing.JButton category7Button6;
+    private javax.swing.JButton category7Button7;
+    private javax.swing.JButton category7Button8;
+    private javax.swing.JButton category8Button1;
+    private javax.swing.JButton category8Button2;
+    private javax.swing.JButton category8Button3;
+    private javax.swing.JButton category8Button4;
+    private javax.swing.JButton category8Button5;
+    private javax.swing.JButton category8Button6;
+    private javax.swing.JButton category8Button7;
+    private javax.swing.JButton category8Button8;
+    private javax.swing.JPanel categoryPanel1;
+    private javax.swing.JPanel categoryPanel2;
+    private javax.swing.JPanel categoryPanel3;
+    private javax.swing.JPanel categoryPanel4;
+    private javax.swing.JPanel categoryPanel5;
+    private javax.swing.JPanel categoryPanel6;
+    private javax.swing.JPanel categoryPanel7;
+    private javax.swing.JPanel categoryPanel8;
+    private javax.swing.JLabel categoy1Label;
+    private javax.swing.JLabel categoyLabel1;
+    private javax.swing.JLabel categoyLabel2;
+    private javax.swing.JLabel categoyLabel3;
+    private javax.swing.JLabel categoyLabel4;
+    private javax.swing.JLabel categoyLabel5;
+    private javax.swing.JLabel categoyLabel6;
+    private javax.swing.JLabel categoyLabel7;
+    private javax.swing.JScrollPane jScrollPane2;
+    private javax.swing.JList<String> orderJList;
+    private javax.swing.JLabel orderTotalLabel;
+    private javax.swing.JLabel orderTotalLabel1;
+    private javax.swing.JButton placeOrderButton;
+    private javax.swing.JButton removeItemButton;
+    // End of variables declaration//GEN-END:variables
+}
