+++ Menu_editor/src/user_interface/StatsUI.java
@@ -1,0 +1,227 @@
+package user_interface;
+
+import middle_end.StatsPage;
+
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+/**
+ *
+ * @author Dylan Girard
+ */
+public class StatsUI extends javax.swing.JFrame {
+
+    /**
+     * Creates new form StatsUI
+     */
+   // private final StatsPage statsPage;
+    private String monthlyProfit;
+    public StatsUI() {
+        //this.statsPage = new StatsPage();
+        initComponents();
+        setVisible(true);
+    }
+    
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        jLabel1 = new javax.swing.JLabel();
+        jLabel2 = new javax.swing.JLabel();
+        jLabel3 = new javax.swing.JLabel();
+        jLabel4 = new javax.swing.JLabel();
+        jLabel5 = new javax.swing.JLabel();
+        jLabel6 = new javax.swing.JLabel();
+        jLabel7 = new javax.swing.JLabel();
+        monthlyProfitLabel = new javax.swing.JLabel();
+        monthlyPrepTimeLabel = new javax.swing.JLabel();
+        dishMostNameLabel = new javax.swing.JLabel();
+        dishMostCountLabel = new javax.swing.JLabel();
+        dishLeastNameLabel = new javax.swing.JLabel();
+        dishLeastLabel = new javax.swing.JLabel();
+
+        setDefaultCloseOperation(javax.swing.WindowConstants.DISPOSE_ON_CLOSE);
+
+        jLabel1.setText("Statistics View");
+
+        jLabel2.setText("Total Monthly Profit");
+
+        jLabel3.setText("Dish Ordered Most Frequently");
+
+        jLabel4.setText("Dish Ordered Least Frequently");
+
+        jLabel5.setText("Number of Times Dish Was Ordered");
+
+        jLabel6.setText("Number of Times Dish Was Ordered");
+
+        jLabel7.setText("Total Monthly Preparation Time");
+
+        StatsPage statsPage = new StatsPage();
+        statsPage.setMonthlyProfit(123);
+        String monthlyProfitText = "" + statsPage.getMonthlyProfit();
+        monthlyProfitLabel.setText(monthlyProfitText);
+
+        statsPage.setMonthlyPrepTime(100);
+        String monthlyPrepTimeText = "" + statsPage.getMonthlyPrepTime();
+        monthlyPrepTimeLabel.setText(monthlyPrepTimeText);
+
+        statsPage.setMostDishesName("Cheeseburger");
+        String mostDishesNameText = "" + statsPage.getMostDishesName();
+        dishMostNameLabel.setText(mostDishesNameText);
+
+        statsPage.setMostDishes(99);
+        String mostDishesText = "" + statsPage.getMostDishes();
+        dishMostCountLabel.setText(mostDishesText);
+
+        statsPage.setLeastDishesName("Asparagus");
+        String leastDishesNameText = "" + statsPage.getLeastDishesName();
+        dishLeastNameLabel.setText(leastDishesNameText);
+
+        statsPage.setLeastDishes(12);
+        String leastDishesText = "" + statsPage.getLeastDishes();
+        dishLeastLabel.setText(leastDishesText);
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
+        getContentPane().setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(layout.createSequentialGroup()
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                            .addComponent(jLabel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                            .addComponent(jLabel3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                            .addComponent(jLabel4, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                            .addComponent(jLabel5, javax.swing.GroupLayout.DEFAULT_SIZE, 310, Short.MAX_VALUE)
+                            .addComponent(jLabel6, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                            .addComponent(jLabel7, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                        .addGap(139, 139, 139)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(monthlyProfitLabel, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                            .addComponent(monthlyPrepTimeLabel, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                            .addComponent(dishMostNameLabel, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                            .addComponent(dishMostCountLabel, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                            .addComponent(dishLeastNameLabel, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                            .addComponent(dishLeastLabel, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                        .addContainerGap())
+                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
+                        .addGap(0, 328, Short.MAX_VALUE)
+                        .addComponent(jLabel1)
+                        .addGap(333, 333, 333))))
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jLabel1)
+                .addGap(18, 18, 18)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel2)
+                    .addComponent(monthlyProfitLabel))
+                .addGap(18, 18, 18)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel7)
+                    .addComponent(monthlyPrepTimeLabel))
+                .addGap(18, 18, 18)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel3)
+                    .addComponent(dishMostNameLabel))
+                .addGap(18, 18, 18)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel5)
+                    .addComponent(dishMostCountLabel))
+                .addGap(18, 18, 18)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel4)
+                    .addComponent(dishLeastNameLabel))
+                .addGap(18, 18, 18)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel6)
+                    .addComponent(dishLeastLabel))
+                .addContainerGap(87, Short.MAX_VALUE))
+        );
+
+        pack();
+    }// </editor-fold>//GEN-END:initComponents
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String args[]) {
+        /* Set the Nimbus look and feel */
+        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
+        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
+         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
+         */
+        try {
+            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
+                if ("Nimbus".equals(info.getName())) {
+                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
+                    break;
+                }
+            }
+        } catch (ClassNotFoundException ex) {
+            java.util.logging.Logger.getLogger(StatsUI.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (InstantiationException ex) {
+            java.util.logging.Logger.getLogger(StatsUI.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (IllegalAccessException ex) {
+            java.util.logging.Logger.getLogger(StatsUI.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
+            java.util.logging.Logger.getLogger(StatsUI.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        }
+        //</editor-fold>
+
+        /* Create and display the form */
+        java.awt.EventQueue.invokeLater(new Runnable() {
+            @Override
+            public void run() {
+                new StatsUI().setVisible(true);
+            }
+        });
+    }
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JLabel dishLeastLabel;
+    private javax.swing.JLabel dishLeastNameLabel;
+    private javax.swing.JLabel dishMostCountLabel;
+    private javax.swing.JLabel dishMostNameLabel;
+    private javax.swing.JLabel jLabel1;
+    private javax.swing.JLabel jLabel2;
+    private javax.swing.JLabel jLabel3;
+    private javax.swing.JLabel jLabel4;
+    private javax.swing.JLabel jLabel5;
+    private javax.swing.JLabel jLabel6;
+    private javax.swing.JLabel jLabel7;
+    private javax.swing.JLabel monthlyPrepTimeLabel;
+    private javax.swing.JLabel monthlyProfitLabel;
+    // End of variables declaration//GEN-END:variables
+//statsPage.setMonthlyProfit(9999);
+       /* private String monthlyProfitText = "" + statsPage.getMonthlyProfit();
+        
+        monthlyProfitText = "" + statsPage.getMonthlyProfit();
+        monthlyProfit1.setText(monthlyProfitText);
+
+        monthlyPrepTimeText = "" + statsPage.getMonthlyPrepTime();
+        monthlyPrepTime1.setText(monthlyPrepTimeText);
+
+        mostDishesNameText = "" + statsPage.getMostDishesName();
+        mostDishesName1.setText(mostDishesNameText);
+
+        mostDishesOrderedText = "" + statsPage.getMostDishes();
+        mostDishesOrdered1.setText(mostDishesOrderedText);
+
+        leastDishesNameText = "" + statsPage.getLeastDishesName();
+        leastDishesName1.setText(leastDishesNameText);
+
+        leastDishesOrderedText = "" + statsPage.getLeastDishes();
+        leastDishesOrdered1.setText(leastDishesOrderedText);*/
+}
